Alguma coisa ai man

+++ codes/primitivas.h
@@ -22,7 +22,7 @@ enum primitivas
 
 //void line_circle(int x_centro, int y_centro, int tamanho, pixel ***ptr_pixels);
 void circle_line(int x, int y, int decisao, pixel ***ptr_pixels);
-void fill_spread(int X, int Y, imagem *ptr_desenho, pixel ***ptr_pixels);
+void fill_spread(unsigned short X, unsigned short Y, imagem *ptr_desenho, pixel ***ptr_pixels);
 void paint_pixels(int eixo_x, int eixo_y, pixel ***ptr_pixels);
 void line(int x_final, int y_final, int x_inicial, int y_inicial, pixel ***ptr_pixels);
 void line_dx(int x_final, int y_final, int x_inicial, int y_inicial, int contador, pixel ***ptr_pixels);
